src/components/chat/requirements.txt

flask==2.3.2 
openai==0.27.8 
pyspellchecker==0.7.2 
sentence-transformers==2.2.2 
transformers==4.30.2 
tokenizers==0.13.3 
faiss-cpu
huggingface_hub==0.14.1